refactored api
