+++ client/src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 
-import authReducer from "../pages/Home/authSlice";
+import authReducer from "../pages/Login/authSlice";
 
 export const store = configureStore({
   reducer: {
